# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Аналитика статистических показателей хоккейных игроков",
        "title": "HockeyAnalytics API",
        "contact": {},
        "version": "1.5"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/analytics": {
            "get": {
                "description": "Возвращает баллы моделей по статистике игроков",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аналитика"
                ],
                "summary": "Проанализировать игроков",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.PlayerAnalyticsResponse"
                            }
                        }
                    }
                }
            }
        },
        "/analytics/compare": {
            "get": {
                "description": "Сравнивает двух игроков по статистике и аналитическим моделям",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аналитика"
                ],
                "summary": "Сравнение игроков",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID первого игрока",
                        "name": "player1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID второго игрока",
                        "name": "player2",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Сезон",
                        "name": "season",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ComparisonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analytics/leaders/{stat}": {
            "get": {
                "description": "Возвращает список лучших игроков по выбранному показателю",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аналитика"
                ],
                "summary": "Таблица лидеров",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Статистические показатели",
                        "name": "stat",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Сезон",
                        "name": "season",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Позиция игрока",
                        "name": "position",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Команда",
                        "name": "team",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Количество игроков",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Минимум сыгранных матчей",
                        "name": "min_gp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Сортировка asc/desc",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.LeaderboardResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/analytics/player/{id}/history": {
            "get": {
                "description": "Возвращает статистику и аналитические оценки игрока по сезонам",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Аналитика"
                ],
                "summary": "История игрока",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игрока",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.PlayerHistoryResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/import/csv": {
            "post": {
                "description": "Импортировать статистику csv файлом",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Импорт"
                ],
                "summary": "импорт csv",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV File",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ComparisonResponse": {
            "type": "object",
            "properties": {
                "player1": {
                    "$ref": "#/definitions/handlers.PlayerComparisonData"
                },
                "player2": {
                    "$ref": "#/definitions/handlers.PlayerComparisonData"
                }
            }
        },
        "handlers.LeaderboardResponse": {
            "type": "object",
            "properties": {
                "player": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "season": {
                    "type": "string"
                },
                "team": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "handlers.PlayerAnalyticsResponse": {
            "type": "object",
            "properties": {
                "base_score": {
                    "type": "number"
                },
                "context_score": {
                    "type": "number"
                },
                "normalized_score": {
                    "type": "number"
                },
                "overall_percentile": {
                    "type": "number"
                },
                "overall_score": {
                    "type": "number"
                },
                "player": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "season": {
                    "type": "string"
                },
                "team": {
                    "type": "string"
                }
            }
        },
        "handlers.PlayerComparisonData": {
            "type": "object",
            "properties": {
                "assists": {
                    "type": "integer"
                },
                "base_score": {
                    "type": "number"
                },
                "blocks": {
                    "type": "integer"
                },
                "context_score": {
                    "type": "number"
                },
                "goals": {
                    "type": "integer"
                },
                "hits": {
                    "type": "integer"
                },
                "normalized_score": {
                    "type": "number"
                },
                "overall_score": {
                    "type": "number"
                },
                "player": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "position": {
                    "type": "string"
                },
                "season": {
                    "type": "string"
                },
                "team": {
                    "type": "string"
                }
            }
        },
        "handlers.PlayerHistoryResponse": {
            "type": "object",
            "properties": {
                "base_score": {
                    "type": "number"
                },
                "context_score": {
                    "type": "number"
                },
                "normalized_score": {
                    "type": "number"
                },
                "overall_score": {
                    "type": "number"
                },
                "season": {
                    "type": "string"
                }
            }
        }
    }
}